Notify cart item count after ATC; Change data type to int;

+++ features/transaction/atc_common/src/main/java/com/tokopedia/atc_common/data/model/response/DataResponse.kt
@@ -14,7 +14,7 @@ data class DataResponse(
 
         @SerializedName("cart_id")
         @Expose
-        val cartId: Long = 0,
+        val cartId: Int = 0,
 
         @SerializedName("product_id")
         @Expose
